agri-app/requirements.txt
tensorflow==2.12.0
keras==2.12.0
streamlit==1.26.0
blinker==1.6.2
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
fonttools==4.42.1
graphviz==0.20.1
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.23.0
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
plotly==5.17.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
tenacity==8.2.3
threadpoolctl==3.2.0
tzdata==2023.3
Werkzeug==2.3.7